+++ core/src/main/java/com/shatteredpixel/yasd/items/wands/WandOfLivingEarth.java
@@ -59,7 +59,7 @@ public class WandOfLivingEarth extends DamageWand {
 	}
 	
 	@Override
-	public int max(int lvl) {
+	public float max(float lvl) {
 		return 6 + 2*lvl;
 	}
